Add a trait type - but still working out the concepts
